Move ndc-client into ndc-test
