+++ README.md
@@ -6,7 +6,6 @@ This repository contains the specification for Hasura's Native Data Connectors (
 
 - [Specification](./specification)
 - [Rust models](./ndc-models)
-- [Rust client](./ndc-client)
 - [Reference implementation](./ndc-reference)
 - [Test runner](./ndc-test)
