--- ndc-client/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "ndc-client"
-description = "Client library for the Hasura NDC specification"
-version = "0.1.1"
-edition = "2021"
-
-[features]
-default = ["native-tls"]
-native-tls = ["reqwest/native-tls"]
-rustls = ["reqwest/rustls"]
-
-[dependencies]
-ndc-models = { path = "../ndc-models" }
-
-async-trait = "^0.1.74"
-opentelemetry = "^0.22.0"
-reqwest = { version = "^0.11", features = ["json", "multipart"], default-features = false }
-serde = "^1.0"
-serde_json = { version = "^1.0", features = ["preserve_order"] }
-url = "^2.4"
