--- ndc-client/README.md
@@ -1,3 +1,0 @@
-# NDC Rust Client
-
-This crate contains the API client code for working with connectors.
