--- ndc-client/src/lib.rs
@@ -1,5 +1,0 @@
-pub mod apis;
-
-/// `ndc_models` is exposed here for backwards compatibility, because it used to
-/// belong to this crate.
-pub use ndc_models as models;
