+++ ndc-test/Cargo.toml
@@ -4,13 +4,8 @@ description = "A tool to verify that a data connector is somewhat compatible wit
 version = "0.1.1"
 edition = "2021"
 
-[features]
-default = ["native-tls"]
-native-tls = ["ndc-client/native-tls", "reqwest/native-tls"]
-rustls = ["ndc-client/rustls", "reqwest/rustls"]
-
 [dependencies]
-ndc-client = { path = "../ndc-client" }
+ndc-models = { path = "../ndc-models" }
 
 async-trait = "^0.1.74"
 clap = { version = "^4", features = ["derive"] }
@@ -20,6 +15,7 @@ rand = { version = "^0.8.5", features = ["small_rng"] }
 reqwest = { version = "^0.11", features = ["json", "multipart"], default-features = false }
 semver = "^1.0.20"
 serde = "^1.0.189"
-serde_json = "^1.0.107"
+serde_json = { version = "^1.0.107", features = ["preserve_order"] }
 thiserror = "^1.0.50"
 tokio = { version = "^1.33.0", features = ["macros", "rt-multi-thread", "parking_lot"] }
+url = "^2.4"
