+++ ndc-test/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod client;
 pub mod configuration;
 pub mod connector;
 pub mod error;
@@ -17,9 +18,7 @@ use colorful::Colorful;
 use connector::Connector;
 use error::Error;
 
-use ndc_client::apis::configuration::Configuration;
-use ndc_client::apis::default_api as api;
-use ndc_client::models::{self};
+use ndc_models as models;
 
 use error::Result;
 
@@ -30,21 +29,21 @@ use serde::{Deserialize, Serialize};
 use snapshot::{snapshot_test, SnapshottingConnector};
 
 #[async_trait(?Send)]
-impl Connector for Configuration {
+impl Connector for client::Configuration {
     async fn get_capabilities(&self) -> Result<models::CapabilitiesResponse> {
-        Ok(api::capabilities_get(self).await?)
+        Ok(client::capabilities_get(self).await?)
     }
 
     async fn get_schema(&self) -> Result<models::SchemaResponse> {
-        Ok(api::schema_get(self).await?)
+        Ok(client::schema_get(self).await?)
     }
 
     async fn query(&self, request: models::QueryRequest) -> Result<models::QueryResponse> {
-        Ok(api::query_post(self, request).await?)
+        Ok(client::query_post(self, request).await?)
     }
 
     async fn mutation(&self, request: models::MutationRequest) -> Result<models::MutationResponse> {
-        Ok(api::mutation_post(self, request).await?)
+        Ok(client::mutation_post(self, request).await?)
     }
 }
