+++ ndc-test/src/main.rs
@@ -1,14 +1,13 @@
 use std::{path::PathBuf, process::exit};
 
 use clap::{Parser, Subcommand};
-use ndc_client::apis::configuration::Configuration;
 use ndc_test::{
     benchmark_report,
+    client::Configuration,
     configuration::{TestConfiguration, TestGenerationConfiguration},
     reporter::{ConsoleReporter, TestResults},
     ReportConfiguration,
 };
-use reqwest::header::HeaderMap;
 
 #[derive(Parser)]
 #[command(author, version, about, long_about = None)]
@@ -127,9 +126,7 @@ async fn main() {
 
             let configuration = Configuration {
                 base_path: endpoint,
-                user_agent: None,
                 client: reqwest::Client::new(),
-                headers: HeaderMap::new(),
             };
 
             let mut reporter = (ConsoleReporter::default(), TestResults::default());
@@ -149,9 +146,7 @@ async fn main() {
         } => {
             let configuration = Configuration {
                 base_path: endpoint,
-                user_agent: None,
                 client: reqwest::Client::new(),
-                headers: HeaderMap::new(),
             };
 
             let mut reporter = (ConsoleReporter::default(), TestResults::default());
@@ -174,9 +169,7 @@ async fn main() {
         } => {
             let configuration = Configuration {
                 base_path: endpoint,
-                user_agent: None,
                 client: reqwest::Client::new(),
-                headers: HeaderMap::new(),
             };
 
             let mut reporter = (ConsoleReporter::default(), TestResults::default());
