+++ ndc-test/src/snapshot.rs
@@ -6,7 +6,7 @@ use std::{
 };
 
 use async_trait::async_trait;
-use ndc_client::models;
+use ndc_models as models;
 
 use crate::{
     connector::Connector,
